# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BorderCityBoxing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coaches/status:
    get:
      operationId: getCoachStatus
      tags: [coaches]
      summary: Check if the current user is a coach
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: Coach status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /coaches/register:
    post:
      operationId: registerAsCoach
      tags: [coaches]
      summary: Register as a coach using a coach code
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachRegisterRequest"
      responses:
        "201":
          description: Registered as coach
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachStatus"
        "400":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /classes/{classId}/coach-signin:
    post:
      operationId: coachSignInToClass
      tags: [coaches]
      summary: Coach signs in to a class
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Coach signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachSignin"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not a coach
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Already signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: coachSignOutFromClass
      tags: [coaches]
      summary: Coach signs out from a class
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutSuccess"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not a coach
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /classes/{classId}/roster:
    get:
      operationId: getClassRoster
      tags: [coaches]
      summary: Get the roster for a class (coaches only)
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class roster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassRoster"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not a coach
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  parameters:
...
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CoachStatus:
      type: object
      required: [isCoach]
      properties:
        isCoach:
          type: boolean
    CoachRegisterRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
    CoachSignin:
      type: object
      required: [id, classId, userId, signinAt]
      properties:
        id:
          type: integer
        classId:
          type: integer
        userId:
          type: string
        signinAt:
          type: string
          format: date-time
    RosterMember:
      type: object
      required: [id, firstName, lastName, bookedAt, isCheckedIn]
      properties:
        id:
          type: integer
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        bookedAt:
          type: string
          format: date-time
        isCheckedIn:
          type: boolean
    ClassRoster:
      type: object
      required: [members, totalAttending, checkedInCount, coachesSignedIn]
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/RosterMember"
        totalAttending:
          type: integer
        checkedInCount:
          type: integer
        coachesSignedIn:
          type: array
          items:
            type: string
...